Improve destructor correctness
Destructors of classes with virtual functions should be marked virtual
or non-virtual and non-public. As the CAN buffers are not deleted
through a pointer to the base buffer class we make the base buffer class
destructor protected and non virtual but make the derived classes have
virtual destructors.

Use cancelAndDelete instead of raw deletes in the destructors.